Split init tests
